requirements.txt
python-telegram-bot==21.6
yt-dlp==2024.12.13
flask==3.1.0
